+++ package/lib/src/ui/components/input_fields/text_input_field_container.dart
@@ -1,20 +1,26 @@
 import 'package:dhis2_flutter_ui/src/ui/models/input_field.dart';
+import 'package:dhis2_flutter_ui/src/ui/utils/validator_util.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
 class TextInputFieldContainer extends StatefulWidget {
-  const TextInputFieldContainer({
-    Key? key,
-    required this.inputField,
-    this.onInputValueChange,
-    this.inputValue,
-    this.showInputCheckedIcon = true,
-  }) : super(key: key);
+  const TextInputFieldContainer(
+      {Key? key,
+      required this.inputField,
+      this.onInputValueChange,
+      this.inputValue,
+      this.showInputCheckedIcon = true,
+      this.validators,
+      this.inputFormatters})
+      : super(key: key);
 
   final InputField inputField;
   final Function? onInputValueChange;
   final String? inputValue;
   final bool showInputCheckedIcon;
+  final List<FormValidator>? validators;
+  final List<TextInputFormatter>? inputFormatters;
 
   @override
   _TextInputFieldContainerState createState() =>
@@ -88,6 +94,7 @@ class _TextInputFieldContainerState extends State<TextInputFieldContainer> {
                   )
                 : textController,
             onChanged: onValueChange,
+            inputFormatters: widget.inputFormatters,
             maxLines: widget.inputField.valueType == 'LONG_TEXT' ? null : 1,
             keyboardType: TextInputType.text,
             textInputAction: TextInputAction.next,