+++ package/CHANGELOG.md
@@ -1,3 +1,10 @@
+
+
+## 1.0.8
+
+- Add support for Validators for TEXT type InputField
+- Add support for InputMask for TEXT type InputField
+
 ## 1.0.7
 
 - Add support for dynamic changes of option list on selection inputs